fetch balance from API, update contract addresses
Ideally we would always read the right contract address from the BE, but it's not that straightforward because then we have to deal with loading states while the data is being retrieved. Thus, for now, I've just updated the hardcoded addresses here. Once we shift to the actual token (and update the addresses once more) this isn't going to be a big deal anyways (I supose).